refactor(workspace-runtime): extract session lifecycle owner
