+++ scripts/ci/module-impact.json
@@ -31,7 +31,8 @@
         "src/renderer/src/lib/acp/workspace-events.ts",
         "src/renderer/src/lib/acp/workspace-runtime-event-owner.ts",
         "src/renderer/src/lib/acp/workspace-runtime-prompt-preparation-owner.ts",
-        "src/renderer/src/lib/acp/workspace-runtime-command-owner.ts"
+        "src/renderer/src/lib/acp/workspace-runtime-command-owner.ts",
+        "src/renderer/src/lib/acp/workspace-runtime-session-lifecycle-owner.ts"
       ],
       "interfacePaths": ["src/renderer/src/lib/acp/useWorkspaceAgentRuntime.ts"],
       "consumerModules": ["workspace_page"],
